LayoutTests/webaudio/convolver-channels-expected.txt

Test Supported Number of Channels for ConvolverNode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ConvolverNode with buffer of 1 channels did not throw exception.
PASS ConvolverNode with buffer of 2 channels did not throw exception.
PASS ConvolverNode with buffer of 3 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 4 channels did not throw exception.
PASS ConvolverNode with buffer of 5 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 6 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 7 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 8 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 9 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 10 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 11 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 12 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 13 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 14 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 15 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 16 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 17 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 18 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 19 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 20 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 21 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 22 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 23 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 24 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 25 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 26 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 27 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 28 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 29 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 30 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 31 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS ConvolverNode with buffer of 32 channels threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS successfullyParsed is true

TEST COMPLETE
